Fixes #2426 Build on GitHub Actions instead of AppVeyor
