--- appveyor.yml
@@ -1,64 +1,0 @@
-configuration: Debug
-image: Visual Studio 2022
-
-environment:
-  app_version: '12.1'
-
-version: '$(app_version).{build}'
-
-init: 
-  - cmd: if DEFINED APPVEYOR_PULL_REQUEST_NUMBER (Set build_number=1) ELSE (Set build_number=%APPVEYOR_BUILD_NUMBER%)
-
-dotnet_csproj:
-  patch: true
-  file: '**\*.csproj'
-  version: '$(app_version).$(build_number)'
-  assembly_version: '$(app_version).$(build_number)'
-  file_version: '$(app_version).$(build_number)'
-  informational_version: '$(app_version).$(build_number)'
-
-before_build:
-  - nuget sources add -name bddhelper -source https://ci.appveyor.com/nuget/ospsuite-bddhelper 
-  - nuget sources add -name utility -source https://ci.appveyor.com/nuget/ospsuite-utility 
-  - nuget sources add -name serializer -source https://ci.appveyor.com/nuget/ospsuite-serializer 
-  - nuget sources add -name databinding -source https://ci.appveyor.com/nuget/ospsuite-databinding 
-  - nuget sources add -name texreporting -source https://ci.appveyor.com/nuget/ospsuite-texreporting 
-  - nuget sources add -name databinding-devexpress -source https://ci.appveyor.com/nuget/ospsuite-databinding-devexpress 
-  - nuget sources add -name funcparser -source https://ci.appveyor.com/nuget/ospsuite-funcparser 
-  - nuget sources add -name simmodel -source https://ci.appveyor.com/nuget/ospsuite-simmodel
-  - nuget sources add -name cvodes -source https://ci.appveyor.com/nuget/ospsuite-simmodel-solver-cvodes
-  - nuget restore
-
-build:
-  publish_nuget: true
-  verbosity: minimal
-  publish_nuget_symbols: true
-  project: OSPSuite.Core.sln
-
-skip_branch_with_pr: true
-
-pull_requests: 
-  do_not_increment_build_number: true
-
-nuget: 
-  disable_publish_on_pr: true 
-
-skip_tags: true
-
-skip_commits:
-  files:
-    - '*.md'
-
-branches:
-  only:
-    - develop
-
-test:
-  assemblies:
-    - OSPSuite.Core.Tests.dll
-    - OSPSuite.Core.IntegrationTests.dll
-    - OSPSuite.Presentation.Tests.dll
-    - OSPSuite.Infrastructure.Tests.dll
-    - OSPSuite.R.Tests.dll
-    - OSPSuite.UI.Tests.dll
-
