+++ tests/OSPSuite.Core.IntegrationTests/Data/IntegrationSampleDuplicateHeader.xlsx
