+++ tests/OSPSuite.Core.IntegrationTests/Data/IntegrationSampleMissingColumn.xlsx
