+++ tests/OSPSuite.Core.IntegrationTests/Data/IntegrationSampleWithGroupBy.xlsx
