+++ tests/OSPSuite.Core.IntegrationTests/Data/emptyFile.xlsx
