+++ tests/OSPSuite.Core.IntegrationTests/Data/invalid.xlsx
