+++ tests/OSPSuite.Core.IntegrationTests/Serializers/BuildingBlockXmlSerializerSpecs.cs
@@ -5,6 +5,7 @@ using OSPSuite.Core.Domain.Builder;
 using OSPSuite.Core.Domain.Descriptors;
 using OSPSuite.Core.Domain.Formulas;
 using OSPSuite.Core.Helpers;
+using OSPSuite.Helpers;
 
 namespace OSPSuite.Core.Serializers
 {
