+++ tests/OSPSuite.Core.IntegrationTests/Serializers/DataInfoXmlSerializerSpecs.cs
@@ -1,7 +1,7 @@
 using NUnit.Framework;
 using OSPSuite.Core.Domain;
 using OSPSuite.Core.Domain.Data;
-using OSPSuite.Core.Helpers;
+using OSPSuite.Helpers;
 
 namespace OSPSuite.Core.Serializers
 {
