+++ tests/OSPSuite.Core.IntegrationTests/Serializers/ModelXmlSerializerSpecs.cs
@@ -2,6 +2,7 @@ using NUnit.Framework;
 using OSPSuite.Core.Domain;
 using OSPSuite.Core.Domain.Services;
 using OSPSuite.Core.Helpers;
+using OSPSuite.Helpers;
 
 namespace OSPSuite.Core.Serializers
 {
