+++ tests/OSPSuite.Core.IntegrationTests/Serializers/ModuleXmlSerializerSpecs.cs
@@ -1,7 +1,7 @@
 ﻿using NUnit.Framework;
 using OSPSuite.Core.Domain;
 using OSPSuite.Core.Domain.Builder;
-using OSPSuite.Core.Helpers;
+using OSPSuite.Helpers;
 
 namespace OSPSuite.Core.Serializers
 {
