+++ tests/OSPSuite.Core.IntegrationTests/Serializers/ObjectPathXmlSerializerSpecs.cs
@@ -1,6 +1,7 @@
 using NUnit.Framework;
 using OSPSuite.Core.Domain;
 using OSPSuite.Core.Helpers;
+using OSPSuite.Helpers;
 using OSPSuite.Utility.Container;
 
 namespace OSPSuite.Core.Serializers
