+++ tests/OSPSuite.Core.IntegrationTests/Serializers/SimulationSettingsXmlSerializerSpecs.cs
@@ -3,6 +3,7 @@ using OSPSuite.Core.Domain;
 using OSPSuite.Core.Domain.Builder;
 using OSPSuite.Core.Domain.Formulas;
 using OSPSuite.Core.Helpers;
+using OSPSuite.Helpers;
 using OSPSuite.Utility.Container;
 
 namespace OSPSuite.Core.Serializers
