+++ tests/OSPSuite.Core.IntegrationTests/Serializers/SimulationXmlSerializerSpecs.cs
@@ -1,6 +1,6 @@
 using NUnit.Framework;
 using OSPSuite.Core.Domain.Services;
-using OSPSuite.Core.Helpers;
+using OSPSuite.Helpers;
 
 namespace OSPSuite.Core.Serializers
 {
