+++ tests/OSPSuite.HelpersForTests/OSPSuite.HelpersForTests.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
+    <TargetFramework>net472</TargetFramework>
     <AssemblyName>OSPSuite.HelpersForTests</AssemblyName>
     <RootNamespace>OSPSuite.Helpers</RootNamespace>
     <IsPackable>false</IsPackable>
@@ -9,6 +9,8 @@
 
   <ItemGroup>
     <PackageReference Include="FakeItEasy" Version="7.3.1" />
+    <PackageReference Include="NUnit" Version="3.13.3" />
+    <PackageReference Include="OSPSuite.BDDHelper" Version="4.0.0.1" />
   </ItemGroup>
 
   <ItemGroup>
