+++ tests/OSPSuite.R.Performance/OSPSuite.R.Performance.csproj
@@ -4,11 +4,6 @@
     <IsPackable>false</IsPackable>
     <TargetFramework>net472</TargetFramework>
     <OutputType>Exe</OutputType>
-    <Version>1.0.0</Version>
-    <PackageVersion>1.0.0</PackageVersion>
-    <AssemblyVersion>1.0.0</AssemblyVersion>
-    <FileVersion>1.0.0</FileVersion>
-    <InformationalVersion>1.0.0.0</InformationalVersion>
     <GenerateAssemblyProductAttribute>false</GenerateAssemblyProductAttribute>
     <GenerateAssemblyCompanyAttribute>false</GenerateAssemblyCompanyAttribute>
     <Authors>Open-Systems-Pharmacology</Authors>
