+++ tests/OSPSuite.R.Tests/OSPSuite.R.Tests.csproj
@@ -42,7 +42,6 @@
     <ProjectReference Include="..\..\src\OSPSuite.Core\OSPSuite.Core.csproj" />
     <ProjectReference Include="..\..\src\OSPSuite.Infrastructure\OSPSuite.Infrastructure.csproj" />
     <ProjectReference Include="..\..\src\OSPSuite.R\OSPSuite.R.csproj" />
-    <ProjectReference Include="..\OSPSuite.Core.IntegrationTests\OSPSuite.Core.IntegrationTests.csproj" />
     <ProjectReference Include="..\OSPSuite.HelpersForTests\OSPSuite.HelpersForTests.csproj" />
   </ItemGroup>
